Improves IoTagent deployment
It is so important to add the parammeter

`IOTA_REGISTRY_TYPE=mongodb`

for the IoTagent container in order to make
the resgistered devices persistent!!

I had it wrong setup because I though that must
be the MongoDB container name but it mustn't it's
an internal parameter!! not like `IOTA_MONGO_HOST`
which must be the container hostname. So keep this
in mind.